Created Multithreaded Neural Network

+++ Lua/NES/SQL.lua
@@ -1,42 +1,54 @@
 function Open()
-	emu.opendatabase()
+	if tonumber(forms.getthreadNum())==0 then
+		emu.opendatabase()
+	end
 end
 
 function DummyRow()
-	emu.writecommand([[
- 	PRAGMA read_uncommitted =1;
- 	insert into rewards (done) values (0);	
- 	]])
+	if tonumber(forms.getthreadNum())==0 then
+		emu.writecommand([[
+	 	PRAGMA read_uncommitted =1;
+	 	insert into rewards (done) values (0);	
+	 	]])
+	end
 end
 
 function DummyRowLoad()
-	emu.writecommand([[
- 	PRAGMA read_uncommitted =1;
- 	insert into rewards (done) values (4);	
- 	]])
+	if tonumber(forms.getthreadNum())==0 then
+		emu.writecommand([[
+	 	PRAGMA read_uncommitted =1;
+	 	insert into rewards (done) values (4);	
+	 	]])
+	end
 end
 
 
 function DummyRowDead()
-	emu.writecommand([[
- 	PRAGMA read_uncommitted =1;
- 	insert into rewards (done) values (2);	
- 	]])
+	if tonumber(forms.getthreadNum())==0 then
+		emu.writecommand([[
+	 	PRAGMA read_uncommitted =1;
+	 	insert into rewards (done) values (2);	
+	 	]])
+	end
 end
 function DummyRowEnd()
-	emu.writecommand([[
- 	PRAGMA read_uncommitted =1;
- 	insert into rewards (done) values (3);	
- 	]])
+	if tonumber(forms.getthreadNum())==0 then
+		emu.writecommand([[
+	 	PRAGMA read_uncommitted =1;
+	 	insert into rewards (done) values (3);	
+	 	]])
+	end
 end
 
 
 
 function EraseLastAction()
-	emu.writecommand([[
-		PRAGMA read_uncommitted =1; 
-		delete from rewards where score is NULL;
-	]])
+	if tonumber(forms.getthreadNum())==0 then
+		emu.writecommand([[
+			PRAGMA read_uncommitted =1; 
+			delete from rewards where score is NULL;
+		]])
+	end
 end
 
 
@@ -49,28 +61,32 @@ function GatherSpeciesNum()
 end
 
 function UpdateReward(fitness_value)
-	emu.writecommand([[
-	PRAGMA read_uncommitted =1;
-	update rewards
-	set score=]] .. fitness_value .." ,status= " .. status .." WHERE score is NULL;")
-	
+	if tonumber(forms.getthreadNum())==0 then
+		emu.writecommand([[
+		PRAGMA read_uncommitted =1;
+		update rewards
+		set score=]] .. fitness_value .." ,status= " .. status .." WHERE score is NULL;")
+	end
 end
 
 function UpdateGenes(GeneCollection)
-	emu.updategenetable(GeneCollection)
+	if tonumber(forms.getthreadNum())==0 then
+		emu.updategenetable(GeneCollection)
+	end
 end
 
 function CreateGeneTable()
-	console.writeline(emu.writecommand([[
-	PRAGMA read_uncommitted =1;
-	CREATE TABLE Genes 
-	(
-		Species int(11) NOT NULL, 
-		Genome int(11) NOT NULL, 
-		GenomeNum int(11) NOT NULL, 
-		Gene int(11) NOT NULL, 
-		GeneContent varchar(100) NOT NULL, 
-		PRIMARY KEY (Species, Genome, Gene)
-	);]]))
-
+	if tonumber(forms.getthreadNum())==0 then
+		console.writeline(emu.writecommand([[
+		PRAGMA read_uncommitted =1;
+		CREATE TABLE Genes 
+		(
+			Species int(11) NOT NULL, 
+			Genome int(11) NOT NULL, 
+			GenomeNum int(11) NOT NULL, 
+			Gene int(11) NOT NULL, 
+			GeneContent varchar(100) NOT NULL, 
+			PRIMARY KEY (Species, Genome, Gene)
+		);]]))
+	end
 end
